+++ src/modules/FixToggleTriggerVisibility.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "module/Module.hpp"
+#include <Geode/modify/EffectGameObject.hpp>
+
+class $editorModule(FixToggleTriggerVisibility) {};
+
+class $modify(FTTVEffectGameObject, EffectGameObject) {
+    $registerEditorHooks(FixToggleTriggerVisibility)
+
+    void setOpacity(unsigned char opacity);
+};